refactor: move private files away from root and public into
