+++ packages/magicbell/src/client/client.test.ts
@@ -1,8 +1,8 @@
 import { basename } from 'path';
 
 import { setupMockServer } from '../../test/mock-server';
-import { Resource } from '../resource';
-import { RequestClient as Client } from './request-client';
+import { Client as Client } from './client';
+import { Resource } from './resource';
 
 // @ts-expect-error import.meta.glob does exist in vite!
 const resources: [string, Resource][] = Object.entries(import.meta.glob('./resources/*.ts', { eager: true })).map(
