+++ packages/magicbell/src/client/error.ts
