+++ packages/magicbell/src/client/options.test.ts
