+++ packages/magicbell/src/crypto.ts
@@ -14,3 +14,12 @@ export function createHmac(secret: string, data: UserData | string) {
 
   return crypto.createHmac('sha256', secret).update(data).digest('base64');
 }
+
+export function uuid4() {
+  // TODO: should be upgradable to crypto.randomUUID(), introduced by node v14
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+    const r = (Math.random() * 16) | 0;
+    const v = c === 'x' ? r : (r & 0x3) | 0x8;
+    return v.toString(16);
+  });
+}
