+++ packages/magicbell/src/project-resources/imports.ts
@@ -2,9 +2,9 @@
 
 import { type FromSchema } from 'json-schema-to-ts';
 
-import { Resource } from '../resource';
+import { Resource } from '../client/resource';
+import { type RequestOptions } from '../client/types';
 import * as schemas from '../schemas/imports';
-import { type RequestOptions } from '../types';
 
 type CreateImportsResponse = FromSchema<typeof schemas.CreateImportsResponseSchema>;
 type CreateImportsPayload = FromSchema<typeof schemas.CreateImportsPayloadSchema>;
