+++ packages/magicbell/src/project-resources/metrics.ts
@@ -2,9 +2,9 @@
 
 import { type FromSchema } from 'json-schema-to-ts';
 
-import { Resource } from '../resource';
+import { Resource } from '../client/resource';
+import { type RequestOptions } from '../client/types';
 import * as schemas from '../schemas/metrics';
-import { type RequestOptions } from '../types';
 import { MetricsCategories } from './metrics/categories';
 import { MetricsTopics } from './metrics/topics';
