+++ packages/magicbell/src/project-resources/users.ts
@@ -2,10 +2,10 @@
 
 import { type FromSchema } from 'json-schema-to-ts';
 
-import { type IterablePromise } from '../method';
-import { Resource } from '../resource';
+import { type IterablePromise } from '../client/method';
+import { Resource } from '../client/resource';
+import { type RequestOptions } from '../client/types';
 import * as schemas from '../schemas/users';
-import { type RequestOptions } from '../types';
 import { UsersNotifications } from './users/notifications';
 import { UsersPushSubscriptions } from './users/push-subscriptions';
