+++ packages/magicbell/src/project-resources/users/push-subscriptions.ts
@@ -2,10 +2,10 @@
 
 import { type FromSchema } from 'json-schema-to-ts';
 
-import { type IterablePromise } from '../../method';
-import { Resource } from '../../resource';
+import { type IterablePromise } from '../../client/method';
+import { Resource } from '../../client/resource';
+import { type RequestOptions } from '../../client/types';
 import * as schemas from '../../schemas/users/push-subscriptions';
-import { type RequestOptions } from '../../types';
 
 type ListUsersPushSubscriptionsResponse = FromSchema<typeof schemas.ListUsersPushSubscriptionsResponseSchema>;
 type ListUsersPushSubscriptionsPayload = FromSchema<typeof schemas.ListUsersPushSubscriptionsPayloadSchema>;
